+++ src/main/java/org/biouno/unochoice/ChoiceParameter.java
@@ -135,7 +135,7 @@ public class ChoiceParameter extends AbstractScriptableParameter {
      * @return filter length
      */
     public Integer getFilterLength() {
-        return filterLength == null ? 1 : filterLength;
+        return filterLength == null ? (Integer) 1 : filterLength;
     }
 
     // --- descriptor
